+++ public/sitemap.xml
@@ -8,20 +8,20 @@
                 
                 
                 
-                <lastmod>2021-06-30</lastmod>
+                <lastmod>2021-07-05</lastmod>
                 </url><url><loc>/new-img-fromat</loc>
                 
                 
                 
-                <lastmod>2021-06-30</lastmod>
+                <lastmod>2021-07-05</lastmod>
                 </url><url><loc>/sample1</loc>
                 
                 
                 
-                <lastmod>2021-06-30</lastmod>
+                <lastmod>2021-07-05</lastmod>
                 </url><url><loc>/syntax-highlight</loc>
                 
                 
                 
-                <lastmod>2021-06-30</lastmod>
+                <lastmod>2021-07-05</lastmod>
                 </url></urlset>
