--- src/components/atoms/icons/index.ts
@@ -1,7 +1,0 @@
-import ProfileIcon from './profile-icon'
-import PostIcon from './post-icon'
-
-export default {
-  ProfileIcon,
-  PostIcon,
-}
