--- src/components/atoms/tag/index.ts
@@ -1,3 +1,0 @@
-import Tag from './tag'
-
-export default Tag
