+++ src/components/common-footer/index.ts
@@ -1,0 +1,3 @@
+import CommonFooter from './CommonFooter'
+
+export default CommonFooter
