+++ src/components/common-header/index.ts
@@ -1,0 +1,3 @@
+import CommonHeader from './CommonHeader'
+
+export default CommonHeader
