+++ src/components/icons/index.ts
@@ -1,0 +1,7 @@
+import ProfileIcon from './ProfileIcon'
+import PostIcon from './PostIcon'
+
+export default {
+  ProfileIcon,
+  PostIcon,
+}
