--- src/components/molecules/post-item/index.ts
@@ -1,3 +1,0 @@
-import PostItem from './post-item'
-
-export default PostItem
