--- src/components/organisms/common-footer/index.ts
@@ -1,3 +1,0 @@
-import CommonFooter from './common-footer'
-
-export default CommonFooter
