--- src/components/organisms/common-header/index.ts
@@ -1,3 +1,0 @@
-import CommonHeader from './common-header'
-
-export default CommonHeader
