--- src/components/organisms/post-list/index.ts
@@ -1,3 +1,0 @@
-import PostList from './post-list'
-
-export default PostList
