+++ src/components/post-item/index.ts
@@ -1,0 +1,3 @@
+import PostItem from './PostItem'
+
+export default PostItem
