+++ src/components/post-list/index.ts
@@ -1,0 +1,3 @@
+import PostList from './PostList'
+
+export default PostList
