+++ src/pages/about.tsx
@@ -1,4 +1,4 @@
-import Layout from '~/components/layout'
+import Layout from '~/components/Layout'
 
 const meta = {
   title: 'ABOUT',
