+++ src/pages/articles.tsx
@@ -1,5 +1,5 @@
 import Layout from '~/components/Layout'
-import PostList from '~/components/organisms/post-list'
+import PostList from '~/components/post-list'
 
 const meta = {
   title: 'ARTICLES',
